TickerLoader/requirements.txt
numpy==1.18.1
pandas==1.0.3
mysql-connector-python==8.0.18
SQLAlchemy==1.3.16
alpha_vantage==2.1.2